// src/app/core/models/registro-partida.model.ts
export interface RegistroPartida {
    apMaterno: string;
    apPaterno: string;
    direccion: string;
    estado: string;
    libro: string;
    nombre: string;
    numeroDocumento?: string;
    numeroPartida: string;
    numeroPlaca?: string;
    oficina: string;
    registro: string;
    tipoDocumento: string;
    zona: string;
    razonSocial?: string;
    propietario?: string;
}